refactor: move dci into a folder module
Implementing the revert handling will make the DCI implementation bigger, moving this module to a folder will allow for a better organization of the files.
