+++ tycho-indexer/src/extractor/dynamic_contract_indexer/mod.rs
@@ -1,0 +1,2 @@
+pub mod cache;
+pub(super) mod dci;
